Update README.md to reflect 1-May-2014 discussion.
On 1-May-2014 we had a discussion of this design in our monthly (first Thursday) MichiPUG meeting. This version of the README.md is my interpretation of that discussion. 
Note: I am not including the architecture discussion in the README.md. I am limiting the README.md to a kind of problem description and avoiding design as best I can.